requirements_test.txt
pytest==3.2.1
pytest-selenium==1.11.4
